Will use logicalDestPort only in FlowClassifier (#4)
* Will use logicalDestPort only in FlowClassifier

* Integration tests pass with dummy sfc driver (not ovs)

* The driver is set in /etc/neutron/neutron.conf under [sfc] section.

* Use the line 'drivers = dummy' instead of 'drivers = ovs'.

* Move OSGiIntegration test to a non-osgi-based test.

* Add back OSGiIntegrationTest and clean up Live Integration Test
