+++ nsfc-plugin/src/main/java/org/osc/controller/nsfc/utils/ArgumentCheckUtil.java
@@ -48,6 +48,4 @@ public class ArgumentCheckUtil {
            throw new IllegalArgumentException(msg);
        }
    }
-
-
 }
